==> 1-testMusicExtraction.ipynb <==
import os
import numpy as np 
from sklearn.model_selection import train_test_split
import pandas as pd
np.random.seed(42)

song_path = "../mp3s-32k/"

artist_list = os.listdir(song_path)

artist_list

train_list = []
test_list = []
validation_list = []

for artist in artist_list:
    path_album = os.path.join(song_path, artist)
    album_artist =  os.listdir(path_album)
    for album in album_artist:
        path_song_album = os.path.join(path_album, album)
        song_album_artist =  os.listdir(path_song_album)
        song_album_artist_path = [os.path.join(path_song_album, song) for song in song_album_artist]
        
        song_album_artist_path= np.array(song_album_artist_path)
        
        np.random.shuffle(song_album_artist_path)
        
        X_train, X_test = train_test_split(song_album_artist_path, train_size=0.90, shuffle=True)
        X_train, X_valition = train_test_split(X_train, train_size=0.90, shuffle=True)
        train_list.append((X_train, artist))
        test_list.append((X_test, artist))
        validation_list.append((X_valition, artist))


train_array = [{"song_path": song, "artist": artist} for songs, artist in train_list for song in songs]
test_array = [{"song_path": song, "artist": artist} for songs, artist in test_list for song in songs]
validation_array = [{"song_path": song, "artist": artist} for songs, artist in validation_list for song in songs]

train_array

train = pd.DataFrame(train_array)
test = pd.DataFrame(test_array)
validation = pd.DataFrame(validation_array)

train.to_csv('./data/train.csv')
test.to_csv('./data/test.csv')
validation.to_csv('./data/validation.csv')

def verify_dataset(data :pd.DataFrame, artist_list: list )-> None:
    print("Verify Dataset")
    for artist in artist_list:
        print(f"{artist} : {len(data[data.artist == artist])} songs")

verify_dataset(train, artist_list)

verify_dataset(test, artist_list)

verify_dataset(validation, artist_list)
