==> 3-spectrogramToCSV.ipynb <==

import os

def spectrogram_to_csv_by_file(path_to_read:str = "", path_to_save:str = "") -> None:
    import os
    from tqdm import tqdm
    import gzip
    import pickle
    import pandas as pd
    try:
        if len(path_to_read)!=0 or len(path_to_save)!=0:
            listFiles = os.listdir(path_to_read)
            pbar = tqdm(listFiles)
            data = []
            for i in pbar:
                file_path = os.path.join(path_to_read, i)
                file_path = file_path.replace('\\', '/')
                with gzip.open(file_path, 'rb') as f:
                    data_file = pickle.load(f)
    
                data.append({
                    'label': data_file['artist_name'],
                    'path': file_path
                    })
            print("Converting data to csv")
            dataFrame = pd.DataFrame(data)
            dataFrame.to_csv(path_to_save)
            print(f"Done file {path_to_save} save")
    except:
        raise("Error please enter data path_to_read and path_to_save")


datasets = ['train', 'test', 'validation']
dataPath = "./dataset"

for i in datasets:
    print(f"Working on {i} dataset.")
    path_to_read = os.path.join(dataPath, i)
    path_to_read = path_to_read.replace('\\', '/')

    path_to_save = os.path.join(dataPath, f"{i}.csv")
    path_to_save = path_to_save.replace('\\', '/')
    spectrogram_to_csv_by_file(path_to_read = path_to_read, path_to_save = path_to_save)
    print(f"{i} finished.")
